+++ jubeatools/formats/jubeat_analyser/memo/load.py
@@ -2,7 +2,6 @@ from collections import ChainMap
 from copy import deepcopy
 from dataclasses import astuple, dataclass
 from decimal import Decimal
-from functools import reduce
 from itertools import product
 from pathlib import Path
 from typing import Any, Dict, Iterator, List, Mapping, Set, Tuple, Union
@@ -350,4 +349,4 @@ def _load_memo_file(lines: List[str]) -> Song:
 def load_memo(path: Path, **kwargs: Any) -> Song:
     files = load_folder(path)
     charts = [_load_memo_file(lines) for _, lines in files.items()]
-    return reduce(Song.merge, charts)
+    return Song.from_monochart_instances(charts)