+++ jubeatools/formats/jubeat_analyser/memo2/load.py
@@ -1,6 +1,5 @@
 from dataclasses import astuple, dataclass
 from decimal import Decimal
-from functools import reduce
 from itertools import product, zip_longest
 from pathlib import Path
 from typing import Any, Dict, Iterator, List, Mapping, Optional, Set, Tuple, Union
@@ -460,4 +459,4 @@ def _load_memo2_file(lines: List[str]) -> Song:
 def load_memo2(path: Path, **kwargs: Any) -> Song:
     files = load_folder(path)
     charts = [_load_memo2_file(lines) for _, lines in files.items()]
-    return reduce(Song.merge, charts)
+    return Song.from_monochart_instances(charts)