+++ jubeatools/formats/konami/jbsq/load.py
@@ -1,4 +1,3 @@
-from functools import reduce
 from pathlib import Path
 from typing import Any, Optional
 
@@ -16,7 +15,7 @@ def load_jbsq(path: Path, *, beat_snap: int = 240, **kwargs: Any) -> song.Song:
         load_jbsq_file(bytes_, path, beat_snap=beat_snap)
         for path, bytes_ in files.items()
     ]
-    return reduce(song.Song.merge, charts)
+    return song.Song.from_monochart_instances(charts)
 
 
 def load_file(path: Path) -> bytes: